Reword mongo connection logging to be less confusing
